removed and renamed some files
