--- frontend/src/components/title/styled-components/Grid.jsx
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export default styled.section`
-    display: grid;
-    grid-template-rows: repeat(2, 1fr);
-    max-height: 114px;
-    padding: 0 80px;
-
-    @media (max-width: 1127px) {
-        padding: 0 40px;
-    }
-
-    @media (max-width: 743px) {
-        display: none;
-    }
-`;
