+++ frontend/src/components/title/styled-components/TitleContainer.jsx
@@ -1,10 +1,17 @@
 import styled from 'styled-components';
-import Flex from '../../styled-components/Flex';
 
-export default styled(Flex)`
-    grid-row: 1 / 2;
+export default styled.div`
+    padding: 0 80px;
+    padding-top: 24px;
+    display: flex;
+    justify-content: center;
+
+    @media (max-width: 1128px) {
+        padding: 0 40px;
+        padding-top: 24px;
+    }
 
     @media (max-width: 743px) {
-        margin-bottom: 10px;
+        padding: 24px 24px 0 24px;
     }
 `;
