--- frontend/src/components/title/styled-components/TitleFlex.jsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export default styled.div`
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    width: 100%;
-    max-width: 1128px;
-`;
